Move cache manipulation to service module
